A start to memsnap
I'm starting at the part that parses the file in /proc/<pid>/maps.
These funtions will build a list of memory regions that the main program
will pull data from.  At the moment, I stuck a main function in
region_list.c so that I can test this part and get it working well
before killing the main function and working on the larger program.