Basic region list functionality
Complete region_list functionality from parse of maps for new
region_lists.  free_region_list is also implemented.  Next up, selective
segment pruning so that segments that aren't rw can be excluded from the
region.  A bit of API shuffling will be required for this, so that's
going to be a future commit.