A working version of memsnap.
Still deficient in many ways.  Hardcoded pid and timer values need to be
replaced.  The error handling is far from what I usually like to see,
the goto is in this case, entirely indicative of bad form.  There's
definitely some missing features, but this is good enough for a
beginning.