Rename handler, call after semaphore is initiated.
Rename the handler from dumpmem, which is not what the handler does, to
alrm_hdlr, which properly reflects that it simply handles the alarm
function and posts to a semaphore.

In addition, the first call is moved to after the semaphore is
initialized, because that's how the code should have been written.