Rename cycle to the more sensible "snapshot", rename -c to -f
Cycles never made as much sense as snapshots, so I thought it would be
better to name them that.  Of course this meant that -c didn't make
sense since -c stood for "cycles" and we already are using -s, so now
I'm using -f for "finish" for that option.  Not the best match, but
should do.  The filenames should make more sense.