Implement option -m
Allow time to be specified in milliseconds.  Overflow behavior is a bit
annoying, but if it exceeds LONG_MAX it will throw an error, which is
probably acceptable.  We set seconds and miliseconds separately and it
appears to work.  Similar behavior is in place for the seconds argument.
-m will overflow approximately a thousand times quicker, but the idea
that someone would be using -m for values that high seems fairly remote.
-s would do better and counting LONG_MAX numbers of seconds seems
unlikely as well.  We can always go higher later if needed...