Implement option -f
Option -f stops after a certain number of snapshots.  The snapshot count
now starts at 1 instead of 0, which means the number the user provides
for the -f option will be the same as the number of the last snapshot.
If it was kept as before, off-by-one issues would likely lead to user
confusion.  An argument of zero is acceptable, but will instruct memsnap
to continue snapshotting the process indefinitely.