Fix read/write loop (nudge by Joe Auricchio)
I originally wrote the read/write loop without error checking, given my
requests were all going to hit even page boundaries, this seemed safe
enough for the first testing of this tool.  Of course, I had forgotten
to go back and implement error checking on the chance read() or write()
did in fact return a partial result.  Thanks to Joe for reminding me to
go back and do this right.

(In other news, I would have done this with splice(), but it was
rejecting my attempts with EBADF and since I didn't see any bugs, I'm
assuming that this is because the kernel treats /proc files strangely
and that is splice()'s way of saying that.  If my assumption is right,
it should probably be documented in the splice(2) manpage.)