Add -g option and problematic and undocumented -s option.
-g globs the segments into a file all jumbled together.  So that memory
is all snapshotted into one single file with no gaps.  -s snapshots each
segment into a sparse file, with memory segments appropriately beginning
at the file offset corresponding to their memory address.
Unfortunately, since many systems don't appear to actually allow sparse
files to span a full range of offsets that are as large as memory, this
code often fails.  The most common error is caught and a warning is
given to the user, warning them that the lseek failed and the feature
isn't going to work very well and indeed that failure is why it is
undocumented.