Gracefully handle processes dying, permission errors, etc
Some race conditions are still likely lurking in some parts of this
code, but some of this stuff is inheriently racey and the reality is
there's a limit of what seems reasonable to do when someone terminates a
pid you're trying to trace out from under you.  We will gracefully deal
with this type of thing for a good portion of the common cases.  We will
also handle invalid pids being specified from the start, pids you aren't
allowed to trace, etc.