Add -c flag, which hackily generates corefiles.
I originally wasn't going to write this feature at all until a later
version because it seemed like it was going to be an annoying pain to
construct the coredumps and output ELF correctly.  It turns out there's
a utility I wasn't aware of called gcore which generates these types of
corefiles and does most of what I wanted to do.  So I figured, why not
go for it?

However, on integrating this utility into memsnap naively, I realized
that it doesn't fit all that well, causes some annoyance and doesn't
work quite like my utility does.  Now this may mean that memsnap really
maybe should work different than it does, but that's beyond the scope of
landing this feature at this particular point in time.  (But maybe it
shouldn't be.)

So basically this is really really hacky.  I debated for awhile over
whether it was a good idea to incorporate this feature at all.  Mostly
because I thought the feature was in fact a useful addition and partly
just because I spent awhile writing this, I decided that perhaps it
should see the light of day.  The feature is documented prominently as
EXPERIMENTAL and I'm still not sure how good I feel about any of this.

I may pull it out again before 1.0, I am not happy with this code.

I am however, happy with ToorCamp, which was where this code was
written.  So for good or bad, you can thank camp for this feature.

For now, push.