Output segment address information
Now for each snapshot a small file containing the address information
for each snapshot is created and filled with information on the address
ranges for each snapshot.  The file is the same name as the rest of the
snapshot files, except instead of _seg*, it is called "_addr", there is
one file per snapshot.

Not enabled for option g, since that's just a big glob.  But it could be
done easily just be removing the if(!OPT_G) statement.