Cleanup: Remove errant subtractions in snprintf()
There were several places where similar code that had been re-used had
destdirlen being subtracted out of the buffer size probably because the
original programmer (yours truly) was an idiot when I first wrote it and
it has kind of just stuck around despite me looking at it a few times
and going "you know, that's uhm, I don't think that, uhm, that is
ridiculous."  But since I was always in the middle of something I didn't
want to remove the subtractions and make it only that way in some
sections, so I kept the same format for all the snprintfs for formatting
path/file names.  So.  So finally I went ahead and just made them all 1)
correct 2) the same.

So I don't think this changes functionality in any way in the least,
though I suppose technically this is one subtraction operation the
processor won't need to run.  But the code definitely is clearer and
doesn't have completely unnecessary subtractions now.