BUGFIX: Segmentation fault in error handling code
When a process can't be ptraced (because say, it finished) some code is
invoked that reports the error.  Printing useful information into the
string that goes to perror requires allocating a buffer and formatting
the string into it using dynamic data.  Turns out... the code was
formatting the error string into the buffer before actually allocating
it, which naturally led to a segfault.

So now the code allocates the buffer, *then* uses it. >_<

Not sure how this bug didn't get caught earlier.  It's pretty obvious.