BUGFIX: Additional dead process detection
Confusingly, a reading from and opening /proc files for a dead process
won't necessarily fail during any of the previous steps (say where one
opens the proc memory file for the process, which completes successfully)
instead, we need to detect an error for the case where the maps file
simply contains nothing.  Now we do and once again, dead process
detection seems to work just a little bit better.

This would not be confusing if this was a race condition type situation,
but this rather seems to come up during most dead process detection
runs.  My normal assumption was that this should all fail sooner.  Hmm.

(Memsnap is still going to miss some dead processes and risk segfaulting
on some of the other possible race conditions.)